moved fetch data functionality to admin dashboard
